tests/wpt/tests/audio-output/selectAudioOutput-sans-user-activation.https.html

<!doctype html>
<head>
<title>Test selectAudioOutput() without user activation</title>
<link rel="help" href="https://w3c.github.io/mediacapture-output/#dom-mediadevices-selectaudiooutput">
</head>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
'use strict';
promise_test(t => {
  const p = navigator.mediaDevices.selectAudioOutput();
  // Race a settled promise to check that the returned promise is already
  // rejected.
  return promise_rejects_dom(
    t, 'InvalidStateError', Promise.race([p, Promise.resolve()]),
    'selectAudioOutput should have returned an already-rejected promise.');
}, `selectAudioOutput() before user activation`);
</script>
